frameworks/deepagents/requirements.txt
deepagents
python-dotenv
langchain-ollama
